refactor(sidebar): share AccountMenu and active-route helper
Extract the duplicated account popover (avatar trigger, header, theme
toggle, logout) into a shared AccountMenu component consumed by both the
app sidebar and the settings rail; the app sidebar passes its Settings
link via children. Drop the duplicate isSettingsItemActive in favor of
the shared isSidebarItemActive from SidebarShell.
